feat(reference): separate abstractions from configuration
Configuration of options has been extracted into
configuration/ module. This allowed to separate
all abstractions from the configuration.

apidom-reference package now exports two types
of configuration - saturated and empty.

Saturated configuration setups all the package components
and wire them into options.

Empty configuration setups empty options.

sideEffects package.json field has been turned on for
the satured configuration.

Subpath exports has been utilized to allow importing
individual components without importing everything.

Refs #2718
