+++ packages/apidom-reference/src/configuration/empty.ts
@@ -1,0 +1,1 @@
+export * from '../index';
