+++ packages/apidom-reference/src/index.ts
@@ -10,38 +10,33 @@ import resolveFn, { resolveApiDOM as resolveApiDOMFn } from './resolve';
 import { readFile as readFileFn } from './resolve/util';
 import dereferenceFn, { dereferenceApiDOM as dereferenceApiDOMFn } from './dereference';
 
+export { url };
+
 export { default as Parser } from './parse/parsers/Parser';
-export { default as ApiDesignSystemsJsonParser } from './parse/parsers/api-design-systems-json';
-export { default as ApiDesignSystemsYamlParser } from './parse/parsers/api-design-systems-yaml';
-export { default as OpenApiJson3_0Parser } from './parse/parsers/openapi-json-3-0';
-export { default as OpenApiYaml3_0Parser } from './parse/parsers/openapi-yaml-3-0';
-export { default as OpenApiJson3_1Parser } from './parse/parsers/openapi-json-3-1';
-export { default as OpenApiYaml3_1Parser } from './parse/parsers/openapi-yaml-3-1';
-export { default as AsyncApiJson2Parser } from './parse/parsers/asyncapi-json-2';
-export { default as AsyncApiYaml2Parser } from './parse/parsers/asyncapi-yaml-2';
-export { default as JsonParser } from './parse/parsers/json';
-export { default as YamlParser } from './parse/parsers/yaml-1-2';
-export { default as BinaryParser } from './parse/parsers/binary/index-node';
 
-export { default as FileResolver } from './resolve/resolvers/file/index-node';
-export { default as HttpResolverAxios } from './resolve/resolvers/http-axios';
-export { default as HttpResolverSwaggerClient } from './resolve/resolvers/http-swagger-client';
-export { default as HttpResolver } from './resolve/resolvers/HttpResolver';
 export { default as Resolver } from './resolve/resolvers/Resolver';
-export { default as ResolveStrategy } from './resolve/strategies/ResolveStrategy';
-export { default as AsyncApi2ResolveStrategy } from './resolve/strategies/asyncapi-2';
-export { default as OpenApi3_0ResolveStrategy } from './resolve/strategies/openapi-3-0';
-export { default as OpenApi3_1ResolveStrategy } from './resolve/strategies/openapi-3-1';
 
-export { default as AsyncApi2DereferenceStrategy } from './dereference/strategies/asyncapi-2';
-export { default as OpenApi3_0DereferenceStrategy } from './dereference/strategies/openapi-3-0';
-export { default as OpenApi3_1DereferenceStrategy } from './dereference/strategies/openapi-3-1';
+export { default as ResolveStrategy } from './resolve/strategies/ResolveStrategy';
 
 export { default as options } from './options';
 export { merge as mergeOptions } from './options/util';
 
 export { default as Reference } from './Reference';
 export { default as ReferenceSet } from './ReferenceSet';
+
+export {
+  DereferenceError,
+  InvalidSelectorError,
+  MaximumDereferenceDepthError,
+  MaximumResolverDepthError,
+  NotImplementedError,
+  ParserError,
+  PluginError,
+  ResolverError,
+  UnmatchedDereferenceStrategyError,
+  UnmatchedResolveStrategyError,
+  UnmatchedResolverError,
+} from './util/errors';
 
 export const readFile = async (uri: string, options = {}): Promise<Buffer> => {
   const mergedOptions = mergeOptions(defaultOptions, options);
@@ -80,17 +75,3 @@ export const dereferenceApiDOM = async <T extends Element>(
   const mergedOptions = mergeOptions(defaultOptions, options);
   return dereferenceApiDOMFn(element, mergedOptions);
 };
-
-export {
-  DereferenceError,
-  InvalidSelectorError,
-  MaximumDereferenceDepthError,
-  MaximumResolverDepthError,
-  NotImplementedError,
-  ParserError,
-  PluginError,
-  ResolverError,
-  UnmatchedDereferenceStrategyError,
-  UnmatchedResolveStrategyError,
-  UnmatchedResolverError,
-} from './util/errors';
